blockchain - voting - added pagination

+++ blockchain/avalanche/voting-poc/admin-back-end/src/main/kotlin/com/example/voting/admin/model/Page.kt
@@ -1,9 +1,11 @@
 package com.example.voting.admin.model
 
+import java.math.BigInteger
+
 data class Page<T>(
     val pageNumber: Int,
     val pageSize: Int,
     val values: List<T>,
-    val elementsCount: Int,
-    val pageCount: Int,
+    val elementsCount: BigInteger,
+    val pageCount: BigInteger,
 )
